backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
PyPDF2==3.0.1
python-docx==1.1.2
google-generativeai==0.8.3
python-dotenv==1.0.1
